requirements.txt
openai>=1.40.0,<2.0.0
langgraph>=0.2.6,<0.3.0
gradio>=4.44.0,<5.0.0
requests>=2.31.0,<3.0.0
numpy>=1.26.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
streamlit>=1.38.0,<2.0.0
plotly>=5.23.0,<6.0.0
